Fail the Windows dev bundle build when an error occurs.
If any part of the dev bundle generation fails, the script should stop.

I was encountering a barely-noticed failure in the Python extraction
when the download tarball was 404-ing (due to my own fault), but the
script proceeded as if the failure hadn't happened, resulting in a
broken dev bundle.